Separate out pinstalldirs for inclusion by PRRTE
Move the pinstalldirs struct to a new pinstalldirs_types.h
file so it can be cleanly included from PRRTE.
