+++ src/mca/pinstalldirs/pinstalldirs.h
@@ -20,45 +20,9 @@
 
 #include "src/mca/base/pmix_base.h"
 #include "src/mca/mca.h"
+#include "src/mca/pinstalldirs/pinstalldirs_types.h"
 
 BEGIN_C_DECLS
-
-/*
- * Most of this file is just for pmix_info.  The only public interface
- * once pmix_init has been called is the pmix_pinstall_dirs structure
- * and the pmix_pinstall_dirs_expand() call */
-struct pmix_pinstall_dirs_t {
-    char *prefix;
-    char *exec_prefix;
-    char *bindir;
-    char *sbindir;
-    char *libexecdir;
-    char *datarootdir;
-    char *datadir;
-    char *sysconfdir;
-    char *sharedstatedir;
-    char *localstatedir;
-    char *libdir;
-    char *includedir;
-    char *infodir;
-    char *mandir;
-
-    /* Rather than using pkg{data,lib,includedir}, use our own
-       pmix{data,lib,includedir}, which is always set to
-       {datadir,libdir,includedir}/pmix.
-
-       Note that these field names match macros set by configure that
-       are used in Makefile.am files.  E.g., project help files are
-       installed into $(pmixdatadir). */
-    char *pmixdatadir;
-    char *pmixlibdir;
-    char *pmixincludedir;
-};
-typedef struct pmix_pinstall_dirs_t pmix_pinstall_dirs_t;
-
-/* Install directories.  Only available after pmix_init() */
-PMIX_EXPORT extern pmix_pinstall_dirs_t pmix_pinstall_dirs;
-
 /**
  * Expand out path variables (such as ${prefix}) in the input string
  * using the current pmix_pinstall_dirs structure */
